Rename contents to components
