+++ meinberlin/apps/dashboard2/__init__.py
@@ -1,9 +1,9 @@
-from .contents import content
 from .components import DashboardComponent
+from .components import components
 from .components.forms import ModuleFormComponent
 from .components.forms import ModuleFormSetComponent
 from .components.forms import ProjectFormComponent
 
-__all__ = ['content', 'DashboardComponent',
+__all__ = ['components', 'DashboardComponent',
            'ModuleFormComponent', 'ModuleFormSetComponent',
            'ProjectFormComponent']
