+++ meinberlin/apps/dashboard2/urls.py
@@ -1,7 +1,7 @@
 from django.conf.urls import include
 from django.conf.urls import url
 
-from . import content
+from . import components
 from . import views
 
 app_name = 'a4dashboard'
@@ -22,5 +22,5 @@ urlpatterns = [
     url(r'^publish/project/(?P<project_slug>[-\w_]+)/$',
         views.ProjectPublishView.as_view(),
         name='project-publish'),
-    url(r'', include(content.get_urls())),
+    url(r'', include(components.get_urls())),
 ]
